Export restartGameBtn
